requirements.txt
paho-mqtt==1.4.0
websocket_client==0.57.0

